[IMP] stock_barcodes_*: Move todo record methods to picking read wizard
