+++ stock_barcodes_elaboration/wizard/__init__.py
@@ -1,1 +1,2 @@
+from . import stock_barcodes_read_picking
 from . import stock_barcodes_read_todo
