+++ stock_barcodes_elaboration/wizard/stock_barcodes_read_picking.py
@@ -1,0 +1,12 @@
+# Copyright 2019 Sergio Teruel <sergio.teruel@tecnativa.com>
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+from odoo import models
+
+
+class WizStockBarcodesReadPicking(models.TransientModel):
+    _inherit = "wiz.stock.barcodes.read.picking"
+
+    def _group_key(self, line):
+        key = super()._group_key(line)
+        key += (line.elaboration_ids,)
+        return key
